Refactor unit tests for start, end & intermediate events with all definitions
