+++ src/model/bpmn/shape/ShapeBpmnEventKind.ts
@@ -38,4 +38,5 @@ export enum ShapeBpmnEventKind {
  *
  * Temporarily used until we support all events
  */
+// TODO When a new kind is supported, uncomment the corresponding line in test/unit/component/parser/json/BpmnJsonParser.event.not.boundary.test.ts
 export const supportedBpmnEventKinds = [ShapeBpmnEventKind.TERMINATE, ShapeBpmnEventKind.TIMER, ShapeBpmnEventKind.MESSAGE, ShapeBpmnEventKind.SIGNAL];
